Add VirtualBox builder.
Update to Packer 0.5.2.

Update to Debian 7.4.0, Ubuntu 12.04.4.

+++ ubuntu-12.04/cleanup.sh
@@ -1,24 +1,26 @@
-#!/bin/bash
-
-# Remove items used for building, since they aren't needed anymore
-apt-get -y remove linux-headers-$(uname -r) build-essential gcc
+# Remove unused packages
 apt-get -y autoremove
 
-#Clean up tmp
-rm -rf /tmp/*
-
 # Removing leftover leases and persistent rules
-echo "cleaning up dhcp leases"
-rm /var/lib/dhcp/*
+if [ -d "/var/lib/dhcp" ]; then
+    echo "cleaning up dhcp leases"
+    rm /var/lib/dhcp/*
+fi
 
 # Make sure Udev doesn't block our network
-# http://6.ptmc.org/?p=164
 echo "cleaning up udev rules"
-rm /etc/udev/rules.d/70-persistent-net.rules
+rm -f /etc/udev/rules.d/70-persistent-net.rules
 mkdir /etc/udev/rules.d/70-persistent-net.rules
 rm -rf /dev/.udev/
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
+rm -f /lib/udev/rules.d/75-persistent-net-generator.rules
 
 echo "Adding a 2 sec delay to the interface up, to make the dhclient happy"
 echo "pre-up sleep 2" >> /etc/network/interfaces
-exit
+
+# Zero out the free space to save space in the final image:
+echo "Zeroing device to make space..."
+dd if=/dev/zero of=/EMPTY bs=1M
+rm -f /EMPTY
+
+#Clean up tmp
+rm -rf /tmp/*
